fix(sdb): Do not check full path if key was specified with `?`
There might be situations where a token has access to a path, but not
subpaths. Checking the full path in those cases will result in a
misleading 403 permission denied error and - if cache flushing
was not disabled - in a token reissuance request.

Since Vault does not behave in a strictly hierachical KV way (as SDB wants
it to), it makes sense to keep the (deprecated in order to make the Vault
SDB syntax more uniform) `?` operator. I would even argue it should
be made mandatory for single-key access to secrets to avoid ambiguity.